Add files via upload
In this project, I have encountered some difficulties with the chart function I defined. When I call the function, I expect the Title of the chart to be displayed as it is a given parameter but that does not happen. I hope someone gives a work around to this. Thank you for your time